Missing dates formatted

+++ src/main/java/com/WDA/bookstore/services/BookService.java
@@ -1,8 +1,10 @@
 package com.WDA.bookstore.services;
 
-import com.WDA.bookstore.dtos.BookDTO;
-import com.WDA.bookstore.exceptions.bookException.LaunchYearItsNotValid;
-import com.WDA.bookstore.exceptions.bookException.PublisherAlreadyRelatedException;
+import com.WDA.bookstore.dtos.outputs.BookOutput;
+import com.WDA.bookstore.exceptions.CantDeleteException;
+import com.WDA.bookstore.exceptions.EntityAlreadyExistsException;
+import com.WDA.bookstore.exceptions.illegalArgumentException;
+import com.WDA.bookstore.exceptions.illegalArgumentException.LaunchYearItsNotValid;
 import com.WDA.bookstore.models.Book;
 import com.WDA.bookstore.models.Publisher;
 import com.WDA.bookstore.repositories.BookRepository;
@@ -29,50 +31,73 @@ public class BookService {
     @Autowired
     private MapperBase mapperBase;
 
-    public List<BookDTO> findAll() {
-        return mapperBase.toList(bookRepository.findAll(), new TypeToken<List<BookDTO>>() {
+    public List<BookOutput> findAll() {
+        return mapperBase.toList(bookRepository.findAll(), new TypeToken<List<BookOutput>>() {
         }.getType());
     }
 
-    public List<BookDTO> findMostRented() {
-        return mapperBase.toList(bookRepository.findAll(), new TypeToken<List<BookDTO>>() {
+    public List<Book> MostRented() {
+        return mapperBase.toList(bookRepository.findMostRented(), new TypeToken<List<BookOutput>>() {
         }.getType());
     }
 
-    public Book create(Book book) {
+    public List<Book> ListOfMostRented() {
+        return mapperBase.toList(bookRepository.findListOfMostRented(), new TypeToken<List<BookOutput>>() {
+        }.getType());
+    }
+
+    public void create(Book book) {
+        book.getId();
         Publisher publisher = book.getPublisher();
-
-        if(publisherAlreadyRelated(publisher.getId(), book.getName())) {
-            throw new PublisherAlreadyRelatedException();
-        } else if(currentYearIsBiggerThanLaunchYear(book.getLaunch())) {
-            publisherRepository.addToRelatedBooks(publisher.getId());
-            return bookRepository.save(book);
-        } else {
-            throw new LaunchYearItsNotValid();
-        }
+        checkIfCanCreateBook(publisher, book);
     }
 
+    @Transactional
     public void update(Book book) {
         book.getId();
-        if(publisherAlreadyRelated(book.getPublisher().getId(), book.getName())) {
-            throw new PublisherAlreadyRelatedException();
-        } else if(currentYearIsBiggerThanLaunchYear(book.getLaunch())) {
-            bookRepository.save(book);
-        } else {
-            throw new LaunchYearItsNotValid();
-        }
+        Publisher publisher = book.getPublisher();
+        checkIfCanUpdateBook(publisher, book);
     }
 
     public void delete(Long id) {
-        bookRepository.deleteById(id);
+        Book book = bookRepository.findById(id).get();
+        Publisher publisher = book.getPublisher();
+        if(book.getTotal_leased() > 0) {
+            throw new CantDeleteException.BookCantBeDeleted();
+        } else {
+            bookRepository.deleteById(id);
+            publisher.setRelated_books(publisher.getRelated_books() - 1);
+            publisherRepository.save(publisher);
+        }
     }
 
     public boolean currentYearIsBiggerThanLaunchYear(Integer launchYear) {
         int currentYear = LocalDate.now().getYear();
-        return launchYear <= currentYear;
+        return launchYear > currentYear;
     }
 
-    public boolean publisherAlreadyRelated(Long editoraId, String nomeLivro) {
-        return bookRepository.existsByPublisherIdAndName(editoraId, nomeLivro);
+    private void checkIfCanUpdateBook(Publisher publisher, Book book) {
+        if(bookRepository.existsByPublisherAndName(publisher, book.getName())) {
+            throw new EntityAlreadyExistsException.PublisherAlreadyRelatedException();
+        } else if(currentYearIsBiggerThanLaunchYear(book.getLaunch())) {
+            throw new illegalArgumentException.LaunchYearItsNotValid();
+        } else {
+            bookRepository.save(book);
+            publisher.setRelated_books(publisher.getRelated_books() + 1);
+            publisherRepository.save(publisher);
+        }
     }
+
+    private void checkIfCanCreateBook(Publisher publisher, Book book) {
+        if(bookRepository.existsByPublisherAndName(publisher, book.getName())) {
+            throw new EntityAlreadyExistsException.PublisherAlreadyRelatedException();
+        } else if(currentYearIsBiggerThanLaunchYear(book.getLaunch())) {
+            throw new LaunchYearItsNotValid();
+        } else {
+            publisher.setRelated_books(publisher.getRelated_books() + 1);
+            publisherRepository.save(publisher);
+            bookRepository.save(book);
+        }
+    }
+
 }
